feat: remove hardhat-gas-reporter, not compatible with smock 2.4.0
hardhat >=2.20.0 is needed to support Cancun features (transient storage).

hardhat 2.20.0 didn't work, so we had to upgrade it above 2.21.0 (when EDR was
included in hardhat)

smock 2.4.0 is needed, because the previous version (2.3.5) was not compatible
with hardhat EDR. smock 2.4.0 assumes the hardhat provider has a `init`
function. That is true when the provider is a LazyInitializationProviderAdapter,
but the package hardhat-gas-reporter wrap that provider with a
EGRDataCollectionProvider and a BackwardsCompatibilityProviderAdapter adapters.
Before 2.4.0, smock checked if the `init` function was available in the
provider, but starting with 2.4.0, it assumes it exists and it calls it
(failing when the BackwardsCompatibilityProviderAdapter is in used)

Workaround: remove the hardhat-gas-reporter package, so the hardhat provider
available in the tests is a LazyInitializationProviderAdapter