feat: add deployments for reference Pendle oracles
Note that I removed proxy deployments for reference oracles on eth
to save some gas – we don't currently use storage in correlated
oracles anyway. However, taking into account capped oracles which
are in the pipeline, we will need to modify the deployment scripts
and redeploy both the live oracles and the reference ones.