feat: add reference oracle conract
Problem: Sometimes we need an on-chain oracle configuration for
monitoring purposes, especially if the production oracle uses an
indirect way to price an asset. There are usually several ways to
price a derivative asset, e.g. for Pendle we could use ptToSy or
ptToAsset rate depending on whether the price information for the
yield token is available on chain and if there is enough liquidity
to make manipulations non-profitable. Since there are different
ways to compute the price, we need a way to set up alternative price
feeds that would:

* use the logic defined in our contracts
* be flexible enough to support underlying assets with no on-chain
  price info
* be managed by a developer wallet so that we don't need a governance
  cycle to update the monitoring

Solution: Implement a ReferenceOracle contract that allows to set
alternative price feeds with a developer multisig and set temporary
prices for assets during the price request for assets with no on-chain
price information.